Attempt to cache package installation before building docs
